requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-shortcuts==1.6
djangorestframework==3.14.0
exceptiongroup==1.1.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.2
python-decouple==3.8
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
